feat: combine edit form and new form
